day18:
Update egui/eframe to 0.26 and fix the moving of window things into viewport command.

day19: Both parts. Part1 was simple (sans parsing) but part2 I got stuck on the ranging for a while. Then as usual for this lots of off by one errors. Good visualization online of the same input really helped nail this down.